pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastcoder"
version = "3.1.0"
description = "Autonomous Software Development Agent — accepts user stories and autonomously writes code, runs tests, deploys, and iterates"
authors = [{ name = "Suman" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "ollama>=0.4.7",
    "gitpython>=3.1.44",
    "aiofiles>=24.1.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
fastcoder = "fastcoder.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["fastcoder"]
omit = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ ==",
]
